requirements.txt
altgraph==0.17.2
APScheduler==3.6.3
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
backports.entry-points-selectable==1.1.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.2
#cssselect2==0.5.0
distlib==0.3.3
# Editable install with no version control (Django==3.2.9)
#-e /Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages
django-embed-video==1.4.1
django-redis-cache==2.1.0
django-rq==2.1.0
django-translation-manager==1.1.1
django-xhtml2pdf==0.0.4
djangorestframework==3.12.4
filelock==3.4.0
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
macholib==1.15.2
mysql-connector-python==8.0.28
oscrypto==1.3.0
Pillow==8.3.2
platformdirs==2.4.0
polib==1.1.1
protobuf==3.19.4
#psycopg2==2.9.1
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyPDF2==1.27.1
PyPDF3==1.0.6
pyTelegramBotAPI==4.1.1
python-bidi==0.4.2
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
#reportlab==3.6.9
requests==2.26.0
rq==1.10.0
six==1.16.0
sqlparse==0.4.2
#svglib==1.2.1
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
tzdata==2021.2.post0
tzlocal==4.0
#uritools==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
#xhtml2pdf==0.2.7
